refactor: fix and change small details
